Split up MixItem.h into Mix{Info,Participant}Item.h
