+++ src/base/QXmppMixParticipantItem.h
@@ -2,43 +2,12 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-#ifndef QXMPPMIXITEM_H
-#define QXMPPMIXITEM_H
+#ifndef QXMPPMIXPARTICIPANTITEM_H
+#define QXMPPMIXPARTICIPANTITEM_H
 
 #include "QXmppPubSubItem.h"
 
-class QXmppMixInfoItemPrivate;
 class QXmppMixParticipantItemPrivate;
-
-class QXMPP_EXPORT QXmppMixInfoItem : public QXmppPubSubItem
-{
-public:
-    QXmppMixInfoItem();
-    QXmppMixInfoItem(const QXmppMixInfoItem &);
-    ~QXmppMixInfoItem();
-
-    QXmppMixInfoItem &operator=(const QXmppMixInfoItem &);
-
-    const QString &name() const;
-    void setName(QString);
-
-    const QString &description() const;
-    void setDescription(QString);
-
-    const QStringList &contactJids() const;
-    void setContactJids(QStringList);
-
-    static bool isItem(const QDomElement &itemElement);
-
-protected:
-    /// \cond
-    void parsePayload(const QDomElement &payloadElement) override;
-    void serializePayload(QXmlStreamWriter *writer) const override;
-    /// \endcond
-
-private:
-    QSharedDataPointer<QXmppMixInfoItemPrivate> d;
-};
 
 class QXMPP_EXPORT QXmppMixParticipantItem : public QXmppPubSubItem
 {
@@ -67,4 +36,4 @@ private:
     QSharedDataPointer<QXmppMixParticipantItemPrivate> d;
 };
 
-#endif  // QXMPPMIXITEM_H
+#endif  // QXMPPMIXPARTICIPANTITEM_H
